pyproject.toml
[tool.poetry]
name = "pymoodef"
version = "0.1.0"
description = "Quiz generation for Moodle in XML."
authors = ["Jose Samos"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9"
pandas = ">=2.2.2"
openpyxl = ">=3.1.2"
pillow = ">=10.3.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.1.2"
pytest-cov = ">=5.0.0"
jupyter = ">=1.0.0"
myst-nb = {version = ">=1.1.0", python = "^3.9"}
sphinx-autoapi = ">=3.0.0"
sphinx-rtd-theme = ">=2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
